backend/pyproject.toml
[project]
name = "supercog-clone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agentic-framework>=0.3.1",
    "aiofiles==24.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.18",
    "aiohttp-cors==0.8.1",
    "aiosignal==1.3.2",
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "anthropic==0.51.0",
    "anyio==4.9.0",
    "argcomplete==3.6.2",
    "attrs==25.3.0",
    "authlib==1.3.1",
    "backoff==2.2.1",
    "beautifulsoup4==4.13.4",
    "blinker==1.9.0",
    "boto3==1.38.9",
    "botocore==1.38.13",
    "browser-use==0.1.40",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "chardet==5.2.0",
    "charset-normalizer==3.4.2",
    "chonkie==1.0.6",
    "click==8.2.0",
    "colorama==0.4.6",
    "coloredlogs==15.0.1",
    "colorful==0.5.6",
    "compressed-rtf==1.0.7",
    "cryptography==44.0.3",
    "dateparser==1.2.1",
    "defusedxml==0.7.1",
    "deprecation==2.1.0",
    "discord-py>=2.5.2",
    "distlib==0.3.9",
    "distro==1.9.0",
    "docx2txt==0.9",
    "duckduckgo-search==8.0.1",
    "ebcdic==1.1.1",
    "et-xmlfile==2.0.0",
    "extract-msg==0.29.0",
    "fastapi[standard]>=0.115.12",
    "fastembed==0.6.1",
    "filelock==3.18.0",
    "filetype==1.2.0",
    "flatbuffers==25.2.10",
    "frozenlist==1.6.0",
    "fsspec==2025.3.2",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "google-ai-generativelanguage==0.6.18",
    "google-api-core==2.24.2",
    "google-auth==2.40.1",
    "google-news-feed==1.1.0",
    "googleapis-common-protos==1.70.0",
    "googlenewsdecoder==0.1.7",
    "greenlet==3.2.2",
    "grpcio==1.71.0",
    "grpcio-health-checking==1.71.0",
    "grpcio-status==1.71.0",
    "grpcio-tools==1.71.0",
    "h11==0.16.0",
    "html2text==2025.4.15",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.31.1",
    "humanfriendly==10.0",
    "icalendar==6.1.3",
    "idna==3.10",
    "imapclient==2.1.0",
    "importlib-metadata==8.7.0",
    "isort>=6.0.1",
    "jinja2==3.1.6",
    "jiter==0.9.0",
    "jmespath==1.0.1",
    "joblib==1.5.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2025.4.1",
    "langchain==0.3.25",
    "langchain-anthropic==0.3.3",
    "langchain-chroma>=0.2.3",
    "langchain-community>=0.3.24",
    "langchain-core==0.3.59",
    "langchain-google-genai==2.1.4",
    "langchain-ollama==0.2.2",
    "langchain-openai==0.3.1",
    "langchain-text-splitters==0.3.8",
    "langgraph>=0.4.3",
    "langsmith==0.3.42",
    "libmagic>=1.0",
    "litellm==1.68.0",
    "loguru==0.7.3",
    "lxml==5.4.0",
    "markdown-it-py==3.0.0",
    "markdownify==0.14.1",
    "markupsafe==3.0.2",
    "mcp==1.7.1",
    "mdurl==0.1.2",
    "mmh3==5.1.0",
    "model2vec==0.5.0",
    "mpmath==1.3.0",
    "msgpack==1.1.0",
    "multidict==6.4.3",
    "narwhals==1.38.2",
    "networkx==3.4.2",
    "numpy==2.2.5",
    "olefile==0.47",
    "ollama==0.4.8",
    "onnxruntime==1.21.1",
    "openai==1.75.0",
    "opencensus==0.11.4",
    "opencensus-context==0.1.3",
    "openpyxl==3.1.5",
    "orjson==3.10.18",
    "packaging==24.1",
    "pandas==2.2.3",
    "pdfminer-six==20191110",
    "pillow==11.0.0",
    "platformdirs==4.3.8",
    "playwright==1.52.0",
    "posthog==4.0.1",
    "primp==0.15.0",
    "prometheus-client==0.21.1",
    "propcache==0.3.1",
    "proto-plus==1.26.1",
    "protobuf==5.29.4",
    "psycopg2-binary==2.9.10",
    "py-rust-stemmers==0.1.5",
    "py-spy==0.4.0",
    "pyarrow==20.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pycryptodome==3.22.0",
    "pydantic==2.11.4",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.9.1",
    "pydeck==0.9.1",
    "pydub==0.25.1",
    "pyee==13.0.0",
    "pygments==2.19.1",
    "pypdf2==3.0.1",
    "pyreadline3==3.5.4",
    "pysocks==1.7.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-magic>=0.4.27",
    "python-multipart==0.0.20",
    "python-pptx==1.0.2",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "ray==2.45.0",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "rich==14.0.0",
    "rpds-py==0.24.0",
    "rsa==4.9.1",
    "s3transfer==0.12.0",
    "safetensors==0.5.3",
    "selectolax==0.3.29",
    "selenium>=4.32.0",
    "setuptools==80.4.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smart-open==7.1.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "soupsieve==2.7",
    "speechrecognition==3.14.2",
    "sqlalchemy==2.0.40",
    "sqlmodel==0.0.24",
    "sse-starlette==2.3.4",
    "starlette==0.46.2",
    "streamlit==1.45.0",
    "sympy==1.13.1",
    "tenacity==9.1.2",
    "textract-supercog==1.6.5.post1",
    "thespian==4.0.1",
    "tiktoken==0.9.0",
    "tokenizers==0.21.1",
    "toml==0.10.2",
    "torch==2.6.0",
    "tornado==6.4.2",
    "tqdm==4.66.5",
    "transformers==4.51.3",
    "typer==0.15.3",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "tzdata==2025.2",
    "tzlocal==5.3.1",
    "urllib3==2.4.0",
    "uvicorn==0.34.2",
    "validators==0.34.0",
    "virtualenv==20.31.2",
    "watchdog==6.0.0",
    "watchfiles==1.0.5",
    "weaviate-client==4.14.1",
    "webdriver-manager>=4.0.2",
    "websockets==15.0.1",
    "win32-setctime==1.2.0",
    "wrapt==1.17.2",
    "xlrd==2.0.1",
    "xlsxwriter==3.2.3",
    "yarl==1.20.0",
    "zipp==3.21.0",
    "zstandard==0.23.0",
]
